Implement TabbableOnlineOverlay component
